sessions/session-05-multi-agents/live-demo/community_management_team/requirements.txt
google-adk>=0.3.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
pydantic>=2.0.0
